estrutura/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
gunicorn>=22.0,<23.0
psycopg[binary]>=3.1,<4.0
